Add enrollments, update courses to use enrollments

+++ site/app/Http/Controllers/CourseController.php
@@ -6,6 +6,7 @@ use App\Course;
 use App\Http\Requests\DestroyCourse;
 use App\Http\Requests\EnableCourse;
 use App\Http\Requests\StoreCourse;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Contracts\Support\Renderable;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
@@ -18,9 +19,12 @@ class CourseController extends Controller
      * Display a listing of the resource.
      *
      * @return Renderable
+     * @throws AuthorizationException
      */
     public function index()
     {
+        $this->authorize('viewAny', Course::class);
+
         $courses = Course::orderBy('created_at', 'desc')
             ->get();
 
@@ -33,9 +37,12 @@ class CourseController extends Controller
      * Display a listing of the resource in the admin panel.
      *
      * @return Renderable
+     * @throws AuthorizationException
      */
     public function manage()
     {
+        $this->authorize('manage', Course::class);
+
         $courses = Course::withTrashed()
             ->orderBy('created_at', 'desc')
             ->paginate(config('const.pagination.per'));
@@ -49,9 +56,12 @@ class CourseController extends Controller
      * Show the form for creating a new resource.
      *
      * @return Renderable
+     * @throws AuthorizationException
      */
     public function create()
     {
+        $this->authorize('create', Course::class);
+
         return view('courses.create');
     }
 
@@ -61,9 +71,12 @@ class CourseController extends Controller
      * @param StoreCourse $request
      *
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function store(StoreCourse $request)
     {
+        $this->authorize('create', Course::class);
+
         // Create new course
         $course = new Course($request->all());
         $course->save();
@@ -78,9 +91,12 @@ class CourseController extends Controller
      * @param Course $course
      *
      * @return Renderable
+     * @throws AuthorizationException
      */
     public function show(Course $course)
     {
+        $this->authorize('view', $course);
+
         return view('courses.show', [
             'course' => $course,
             'cards' => $course->cards
@@ -93,9 +109,12 @@ class CourseController extends Controller
      * @param Course $course
      *
      * @return Renderable
+     * @throws AuthorizationException
      */
     public function edit(Course $course)
     {
+        $this->authorize('update', $course);
+
         return view('courses.edit', [
             'course' => $course
         ]);
@@ -107,9 +126,12 @@ class CourseController extends Controller
      * @param Course $course
      *
      * @return Renderable
+     * @throws AuthorizationException
      */
     public function configure(Course $course)
     {
+        $this->authorize('configure', $course);
+
         return view('courses.configure', [
             'course' => $course
         ]);
@@ -122,10 +144,13 @@ class CourseController extends Controller
      * @param Course $course
      *
      * @return Response
+     * @throws AuthorizationException
      */
     public function update(Request $request, Course $course)
     {
-        //
+        $this->authorize('update', $course);
+
+        // TODO: update specified course
     }
 
     /**
@@ -139,6 +164,8 @@ class CourseController extends Controller
      */
     public function enable(EnableCourse $request, int $id)
     {
+        $this->authorize('enable', Course::class);
+
         $course = Course::withTrashed()->find($id);
 
         $course->restore();
@@ -157,6 +184,8 @@ class CourseController extends Controller
      */
     public function disable(Course $course)
     {
+        $this->authorize('disable', $course);
+
         $course->delete();
 
         return redirect()->back()
@@ -174,6 +203,8 @@ class CourseController extends Controller
      */
     public function destroy(DestroyCourse $request, int $id)
     {
+        $this->authorize('forceDelete', Course::class);
+
         $course = Course::withTrashed()->find($id);
 
         $course->forceDelete();
